Mike:
scsi_host_lookup takes a handle to the scsi host, so you need a 
scsi_host_put to balance it or you leak memory. I thought it was not 
possible to be in iscsi_if_create_cnx and iscsi_if_destroy_snx at the 
same time I released the handle right away (if this is not true the put 
should be moved down).



git-svn-id: svn://svn.berlios.de/open-iscsi@237 d7303112-9cec-0310-bdd2-e83a94d6c2b6