Hannes Reinecke:
I found it really annoying that iscsiadm has to get it's information
from some file (eg /etc/initiatorname.iscsi and /etc/iscsid.conf), when
there is
a) no chance to set the via the commandline
b) this information is already present in the daemon.

So here's a patch which will retrieve this information from the daemon.
And while I've been at it, the initiatorname and config file name can
now the changed from the commandline of the daemon, too.



git-svn-id: svn://svn.berlios.de/open-iscsi@401 d7303112-9cec-0310-bdd2-e83a94d6c2b6