Use global variable to signal end of event loop
Currently, MGMT_IPC_IMMEDIATE_STOP is implemented by returning 1
from mgmt_ipc_handle. This patch introduces a global variable named
leave_event_loop, which signals the event loop that we should exit now.

The underlying motivation of this patch is to unify the IPC handler API,
and get rid of the (sometimes inconsistent, and all but ignored) return
values of this function entirely.

Signed-off-by: olaf.kirch@oracle.com