rm conn fn pointers to ipc fns
There is no need for the connection to carry pointers to
send/recv callbacks. If the discovery code had nicely set some
wrappers then the code would be nice, but it did not. Instead,
for now if the ipc is set, then use it. If not drop down
to read/write/alaram.