use login timer instead of conn_context
When we login userspace thinks we are sort of transitioning
into ffp (calling start_conn and setting params), but the target is
already to go and the network layer could begin sending errors via
the state changed callback. So we do not want to flush all the queued
contexts (even though if we are getting pdus they could be messed up due
to the kernel not yet having settings like digests set). Instead we just
want to delete the specific login timer,  so this uses the login timer
instead of conn context since we need the finer grainularity.