fix early session shutdown
On the initial login the initiator will destroy the session
on a failure. This is so we do not wait forever for that login.
Unfortunately the original code did not do a good job of handling
errors later on. For many errors the code would destroy the session
on instead of retrying.

This patch should modify the code so that we handle those errors a little
nicer and retry.