tmp fix to iface operation execution and use check
The internal match functions assume that the iface is properly bound,
but when we are setting them up this is not the csae. The bound
check is race because another instance of iscsiadm could be setting
that up, although it is not likely so this is fine for now.

The check_for_session_through_iface is also racey, because we could
be logging in a session from another instance of iscsiadm.

Both need some good fixups along with the normal node/iface deletion case.