fix connection state setting from login lib
The login lib does not need to set the state. In the success case
it was setting the connection state to free, when it should have been
logged in. The caller will just set this from now on, so it can
handle the transtions better if it needs them even.