Finish hooking into tool
Instead of having each parser print its output, make it common
in fw_entry.c and make that use the same format and fields as
the rest of the tools.

Kept the ibft dump code, because it may be use for compat tools for
suse. Let me know and I can help on that if you need it. With the hook
in though the format for ibft is the same as normal iscsiadm node info
so you can just reuse your parsing login for that.