add ppc fwparam support
IIs it flora or fauna?  network or disk?  BIOS or OFW?  :)

Here is another stab at providing the ppc64 equivalent of the xseries
iSCSI Boot Firmware Tables functionality in open-iscsi.

When booting from the firmware initiator, the parameters flow up from
the running system v. from the values stored in the OS image.  We have
to manage the transition from kernel running from the disk obtained
from a network device before the normal setup info is available.  The
trick is getting the pertinent data out of the "hardware",  then passing
this info to the tools running on the OS instance.

On certain pseries boxen the system OFW has an initiator.  This
mini-initiator will only use a single target.  It does not use the
multiple initiator list like on xseries.  You can only boot from a
single target.  You can have multiple different targets defined, but
only one may be active at a time.

That being said, only nics that support the load method qualify as
iSCSI bootable devices on pseries.  This does restrict the devices on
this platform to those that support FCODE.

"Features" of the implementation:

- has all the information about the nic in use; address, netmask, mac
  addr, the initiatorname, port, etc., before _anything_ is run from
  the FS save the booted kernel.

- Reads the /proc/device-tree for ppc related device information.

- for retrieving the bootable device, uses the /chosen/bootpath
  property.

- for display purposes, keys off the /aliases property "iscsi-disk*"
  to determine the bootable disks.

- uses the device ordering from firmware (hence the kernel ordering)
  to determine the interface names at boot time.  The interfaces names
  can and do get munged later by runtime tools, re-ordering the names.