grab eh_mutex during host reset
I thought we may not need the eh mutex during host reset, but I was
wrong. When start_session_recovery sets the state to terminate then
drops the session lock. The scsi eh thread could then grab the session
lock see that we are terminating and then return failed to scsi-ml.
scsi-ml's eh then owns the command and will do whatever it wants
with it. But then the iscsi eh thread could grab the session lock
and want to complete the scsi commands that we in the LLD, but
it no longer owns them and kaboom.