Do not return from eh functions if in the middle logging out.
We do not want to return from the scsi eh functions if we
are logging out, because a command could be in the driver,
then we get a logout command. We then must either fail the in
progress command or wait for the session recovery/shutdown code
to clean it up. The latter is much easier and is currently done
for us, so this just stops the scsi eh code from returning before
the command is returned to the scsi layer.