fix shutdown
We were using the device delete sysfs file to remove each device
then logout. Now in 2.6.21 and .22 this will not work because
the sysfs delete file returns immediately and does not wait for
the device removal to complete. This causes a hang if a cache sync
is needed during shutdown.

This patch fixes the shutdown code so that we remove the target
and unbind the session before logging out and shuttdown the session.

Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>