async login support
If we have multiple portals or for something like equalogic multiple
targets, iscsiadm will login to each object sequentially. For setups
with lots of targets or portals this make start up very slow. This
patch modifies login so we login to each portal on each target
asynchronously.

One bug, is that if there is an error logging in the error message
is not easily matched up to the session, but that will be fixed later.