fast logout in userspace.
If there is an error while we are trying to logout, just proceed
to cleanup. This fixes a bug where we did not know what to do and
just hung.

TODO: maybe in the future we can retry the logout N number of times,
but for now, this is a rare case, so we will fix other bugs first.