add SIGTERM iscsid handler and iscsiadm command to stop iscsid
You can now do killall iscsid or iscsiadm -k 0 to stop iscsid.
This should not normally be done, because iscsid is required
for normal initiator behavior. It is mostly for testing, iscsi
boot or system shutdown.

TODO:
- more testing. I am pretty sure the original code is broken. For
exmaple sending a SIGTERM to the log process then the daemon process
will cause semop errors.