print kernel session state
drivers like qla4xxx do not use iscsid's state machine (there
sm is in hw), so iscsid cannot print the state of the session
or connection. In the kernel we export the session state, so
this patch has iscsiadm print it with the other state values
when running iscsiadm -session -P 1.