partial revert of burst setting patch
In this commit 2bfae61e36dd981f28b59b026542f145c70c70c8,
we started to only check the burst lengths in userspace
if immediate data is yes or of initial r2t is no. It turns
out that targets like netapp and iqstore will always check this
value and fail the login, so to avoid that mess this patch
always checks the value. You will end up getting a error
message in cases where you should not but at least the login
always succeeds and we will not have to help the user try
to figure out what their target error message means.