uclibc support
Bug report and Patch from From Natanael Copa:

I'm trying to compile open-iscsi on uclibc to be able to use it on
embedded linux.

I found out that there are 3 functions that are used that are not
implemented in uclibc: getifaddrs, freeifaddrs and versionsort. None of
those are in POSIX afaik.