Use CHAP info from sysfs when restarting iscsid.
If you are starting iscsid with sessions running then you are
probably doing iscsi boot or had to restart due to a iscsid bug.
For iscsi boot, the ibft data is the most up to date so we do
not want to use the node db info. This patch is sort of hacky,
and in the next release we will clean up the iscsi_sysfs code
and export everything we need (the only problem is the the
program run during boot may not set all the values like segment
lengths or timeouts so we may still need this hack).