Fixup debian info
Patch and info from Andrew Moise:

  Okay, patch is attached.  I wound up doing some things differently
than I mentioned -- the trouble with KERNEL_TARGET turned out to just
be because I hadn't installed the needed headers, and I added the
dpkg-divert script as a Makefile target instead of a snippet of code
in the docs that people have to cut and paste.  I also mentioned the
ib_iser situation (which I hope I described accurately).