allow discovery to overwrite records if a session is using it
If a node record exists for a session we should allow discovery to
overwrite it because it causes a regression for users of iscsi_tcp.

The only value we cannot change is iface.transport_name
(the primary binding iface values are checked as part of the stale
record end to end checks), and the iface.transport_name is only changed
by iscsi_discovery and that has its own checks for now.

In the next release we are going to fix up the iface values so transport
name is a binding value, but it is too late for that, so we just do not
want to add any regressions.