convert iscsi_discovery to use iface settings
node.transport_name was replaced with iface.transport_name because
we needed to bind the interface to the driver for qla4xxx and bnx2i.
This converts iscsi_discovery to use the new value.

Note: there was compat support in iscsiadm for both values, but it
was agreed that because of its limited use and hacks needed to support
compat we would just drop it. This will affect distros that do
not carry iscsi_discovery, becuase the primary discovery tool for iser
was that script.