Do not logout session if the record is going to be deleted.
If a session is using a record we do not want to remove the record
because some commands need to check the node.startup and iface settings.

During devel of this release we added code to logout the session then
delete it, but the problem will be if the user just wanted to delete
the record but was still using the session. In this case the devices
are removed and the user will get FS errors.

For now we will just warn and advise the user, so they can manually
log out. In the next release we can add a force flag to logout.