test pattern netdev instead of iface netdev when matching ifaces
When we match netdevs we should check if the pattern is valid/bound. If it
is not then we do not want to try and match the ifaces netdev. We were testing
the iface's netdev instead and that is going to always be set from the
current callers (sysfs iters and from create_node_rec paths).