Do not allow iface setting to be changed in node mode.
This temporarily breaks iscsi_discovery and probably other apps
for distros that wanted to set iface.transport_name. In the
patch to add default a iser iface we broke node.transport_name,
so this means there is no backward compat for that setting
and you must use the new default iser iface.

Before this is released we will add some compat code somewhere and
fix up the check in iscsiadm for iface settings based on Dorn's patch.