Release transport entry when the transport is unloaded
When iscsid detects a new transport (e.g. iscsi_tcp, ib_iser), it
allocates an entry for it in the transport list. If the transport
is unloaded and then reloaded, iscsid would still use the old
transport entry. Therefore, the transport entry must be deleted
when the transport is unloaded.

Signed-off-by: Erez Zilber <erezz@voltaire.com>