libiscsi: fix recv tmo
If the ping tmo is longer than the recv tmo then we could miss a recv
tmo check, because the ping code will set the next timeout for the
ping timeout. This patch has the ping processing code kick off a recv
tmo check when getting a nop in response to our ping.