Only autobind to ifaces with transport = tcp
This just adds back the old behavior as the default where we
would only autobind to ifaces that are using iscsi_tcp for
the transport type.

Autobinding to iser and bnx2i is not working as it is today, because
it is only based on the modules being loaded. It should be based on
if the hardware is properly setup or if the user requested it in
the config or something else.