fix ipv6 login redirect support.
If the target gives us brackets ([]) in the ipv6 name we need
to drop then when passing the ipv6 addr to the network functions.
The discovery and iscsiadm command line code does this already, but
the login redirect code is not, and this results in the login failing
and iscsiadm hanging.