fix dynamic tpgt support.
Older tools allowed you to not set the tpgt when you created
a record manually. If you update a record, iscsiadm will
update the record format for you if it is supportable. In
this case supportable meant that you had to have a tpgt set.
We checked for it in some other cases, but missed it here for
some reasons. This patch fixes that.