iscsi conf: partially revert increase default login max change
45 is too high if the connection is timing out. It is
45 * 15 seconds. This sets the retry limit to 8,
which would make the connection timeout wait 2 minutes.
This is still to few retries for the static ip problem, so
in the next patch we will modify iscsid so it retries
connection errors like this up to login_timeout * login_retry
seconds. This would be 2 minutes which is working for the
bug reporters on the list.