modify initial login retry max
Now the retry count only controls the number of retries
that we attempt when the initial login attempt timesout.
It used to control the retry count for any error, which resulted
in it having to be set really high and it was also difficult to
figure out what to set it to since users (and me) have no idea
how quickly the network layer will return errors in some cases.

The retry count is also used to determine the max amount of time
to wait for the initial login attempt. We do this by:

login retry max * login timeout = max time

This makes it easier for the user to calculate the
max time that they will have to wait.

The default time to wait is now 2 minutes for normal sessions and
7 minutes for root.