iscsid idbm: print out header and tail to seperate records
When printing multiple records like with

iscsiadm -m node -T my_target

(will dump all records for that target), the records are clumped
together and the only way to parse them is by assuming that the
node.name field indicates the beginning of a new record. This just
adds a

"# BEGIN RECORD"

at the begining and a

"# END RECORD"

at the end of a record.