ibft: add support to use iscsi_ibft module and log into all portals found in firmware
This adds the ibft sysfs module (iscsi_ibft) parsing support. It
was original done by Konard, but I have ported it to use the
sysfs.c helpers (add new ones and ported iscsi_sysfs.c too).

This patch also modifies iscsistart and iscsiadm to print/log
into all the portals found in firmware.

It also changes the behavior of

iscsiadm -m discovery -t fw

so that we create db records for what is found. It is not
fully hooked into the iface code, so it will use the
different inititor name, but it will not create a iface for
it (this means that if you have a record for the same portal that
is bound to a iface with a iname you can run into problems).