iscsiadm: have fw discovery handle --op arguments
This had the iscaidm -m discovery -m fw comman handle
the --op arguments.

It also has it handle the -I argument, but it is a little strange,
and it is not fully supported. The problem is that we currently
do not assume that the net info in the firmware means that we
want to bind the session to a specific nic, so with the -I a user
can bind the session and the net info in the firmware is completely
ignored. We probably need an extra flag to indicate if the user
wants iscsiadm to setup ifaces based on the firmware info and
then also bind the session to them.