bind offloaded connection to port
If we are using a offload card, then iface_set_param will match the
iface info to a scsi_host and pass that info down to setup the net
settings of the port (currently we just set the ip address). When we
create the tcp/ip connection by calling ep_connect, we currently just go
by the routing table info.

I think there are two problems with this.

1. Some drivers do not have access to a routing table. Some drivers like
qla4xxx do not even know about other ports.

2. If you have two initiator ports on the same subnet, the user may have
set things up so that session1 was supposed to be run through port1. and
session2 was supposed to be run through port2. It looks like we could
end with both sessions going through one of the ports.

Also how do you edit the routing table for the offload cards? You cannot
use normal net tools like route can you?

3. If we set up hostA in the iface_set_param step, but then the routing
info leads us to hostB, we are stuck.

I did the attached patches to fix this. Basically we just pass down the
scsi host we want to go through.