iscsiadm: fix discovery login command
If you were to run
iscsiadm -m discovery -ip -l
to try and login to a targets found during discovery using the
/etc/iscsi/send_targets/ip settings, and if CHAP was used
then the discovery session login would fail. The reason for
this is because we were using the password_len fields to specify
the password length to try and allow for any chars.

This patch adds those idbm fields back.