iscsid: fix ISCSI_ERR_INVALID_HOST err handling
When handling a iscsi conn error we will flush other
operations begind it in the queue. If one of those operations
is a ISCSI_ERR_INVALID_HOST, then we end up doing iscsi recovery
instead of killing the session, and drop the session cleanup.

This has iscsid add ISCSI_ERR_INVALID_HOST handling at the
head of the queue instead of at the end.