iscsid: fix segfault during session sync up
We have a silly separation of the transport and template,
because we do not need both in many cases and I went overboard
trying to make programs smaller. iscsid is trying to access
the transport template on the transport but it is has not been
set, so iscsid is segfaulting. This adds a set transport template
call. In later patches (the send targets discovery one) we end
up needing the template and transport in the same place, so a clean
up patch for that will make this not needed. But for now so we
do not have the segfault this patch will get us by.