iscsistart option to bring up NICs using configuration in iBFT.
Patch and mail from Alex Zeffertt:

For each target listed, iSCSI Boot Firmware Tables specify which NIC to
use and
how it should be configured.  Until now this information has been
ignored by
open-iscsi.  This patch enables iscsistart to apply the NIC
configuration.

The new command "iscsiadm -n" applies the NIC configuration specified in
the
iBFT for each valid target.

The primary benefit of this is that it allows the initrd to extract
networking
information from the iBFT rather than hard code it.  If the initrd uses
the iBFT
for networking info then when this info is modified via the BIOS it is
not
necessary to rebuild the initrd.

Signed-off-by<alex.zeffertt@eu.citrix.com>