ibft boot: remove be2iscsi
We do not know what be2iscsi is going to support, so this
patch removes the ibft support. If it did do ibft, then
the net driver name would be be2iscsi since the pci info
would match the iscsi offload card. This is different
from cxgb3i/bnx2i, because be2iscsi's iscsi and net functions
are seperate.