offload boot: turn off
No distro is ready for offload boot, so this patch turns it
off. If a distro is ready then just build with OFFLOAD_BOOT_SUPPORTED.

Note to distro people:
- To turn on you will want to include and load the offload driver like
cxgb3i or bnx2i, and for bnx2i you will want to include its
brcm or uip (depends on what version you use) daemon.

That should be it. The iscsi tools should detect and set things
up.