st discovery: fix reopen max handling
reopen max was supposed to tbe the max number of login retries,
but as it is implemented it is the total numnber of login
retries. So if you set it to 1 you would only get one login
attempt, but you probably wanted 1 initial try and 1 retry.
This patch fixes the code so it is the number of retries.

Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>