iscsi tools: fix MaxXmitDataSegmentLength=0 handling
This fixes a regression added with:

Author: Mike Christie <michaelc@cs.wisc.edu>
Date:   Fri Feb 12 22:59:34 2010 -0600
    iscsi tools: add MaxXmitDataSegmentLength param

In that patch if MaxXmitDataSegmentLength=0 then we just
use the target's MaxRecvDataSegmentLength. The bug is that
I did not init the conn->max_xmit_dlength so if the target
does not send a MaxRecvDataSegmentLength. then the xmit_dlength
is zero and we never send any data.