iscsi tools: fix port handling for iscsiadm commands
iscsiadm was resetting the port to 3260 due to a regression
added with c94bb3206d8816f9d13c8d971513202d94cbb9ea. In that
patch the port handling was changed to handle isns's default
port which differred from iscsi's default, but it also ended
up also resetting the port when a valid port was passed in to
iscsiadm.