iscsi scripts: use iscsiadm -k to shutdown daemon
This has the suse and red hat scripts use the iscsiadm
kill command to shutdown the daemon. This is needed
so if the discovery daemon is used then it will be
cleanly stopped.

This does not modify the debian script because I was
not sure what the start-stop stuff was doing and did not
want to add a bug.