iscsi tools/kernel: switch make defaults
With stable kernels and the iscsi kernel modules stablizing, there is
less of a need to distribute kernel modules with the tarball. I also
do not have time to update and maintain ports of the modules for
so many kernels. So this patch changes the build default behavior.

Running "make" use to build the kernel modules and tools. Now, it
will only create the tools. "make install" used to install the
tools and kernel modules and now it only installs the tools.

I am not removing the kernel modules, but I will not be updating them
anymore. They will remain for users of very old kernels. To build
you now need to do

and to install do