iscsid: Fix netdev check
A check for a bound session omitted the subscript for the first byte of
the array, forcing the check to always be true. This means that in the
case of a session not bound to an interface, the ss structure will not
be filled-in if the logging level is set to 0. This will prevent the
DCB priority from being set.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>