iscsi tools: Don't try to bind offload EPs to sockets
If discovery offload is not supported iscsiadm/iscsid
drops down to normal socket based discovery. The
problem is that io.c will try to bind the session
to the socket if the iface has binding info, but
for offload cards we will not be able to bind
based on the hw address. This is because that will
info is private to the iscsi function and not exposed
to the network layer.