iscsi tools: fix bnx2i boot due to MAC mismatch
The problem is current bnx2i boot firmware has 2 modes.

1. hba mode - Will export the iscsi MAC in ibft.
2. non hba mode - Will export the network MAC in ibft.

[In both modes there should be a sysfs link to the network
netdev from the ibft ethernet dir]

For #1 we want to create a session using bnx2i. For #2 we want to use
software iscsi (we can only use software iscsi since we do not know the
MAC for the offload card).

This patch fixes a bug where we would create a iscsi session using the
offload card if the netdev that is found through the ibft sysfs tree was
managed by a net driver that has a iscsi offload driver. With this patch
we now determine if we use the offload card during boot based on if
the /sys/class/iscsi_host/hostX/hwaddres matches the one in iBFT. If
it does, then we know the iBFT MAC is the iscsi one and use the offload
driver.

Note that cxgb*i always uses 1 MAC for both iscsi and offload. So
for that driver we always use offload. To get the netdev we follow
the sysfs ethernet link to the netdev like is done with bnx2i.