Add support for multiple sessions per iface to iscsid
This introduces a new 'multiple' flag to the session_rec_t, which if set
instructs iscsid to bypass the check which would normally prevent
multiple sessions from a single iface record, allowing multiple sessions
to be created.

The commandline syntax to actually set this flag will be introduced in a
later commit.

Signed-off-by: Jim Ramsay <jim_ramsay@dell.com>