iscsi tools: revert commit c440cbe7ba2464f8baadedb55b00754c36773c2c
This reverts:

	commit c440cbe7ba2464f8baadedb55b00754c36773c2c
	Keep startup mode in sync when specified in config file

The normal case is for there to be both the node and conn
startup values in a rec. This is how iscsiadm writes it
out by default. The patch added a memory leak due to the
assumption that there would only be one. Also if they are both
present then the patch ends up not doing anything.