Check all ifaces during discovery even if some timeout
Any errors in the discovery process would previously cause a
short-circuit jump out of the main per-interface loop.  There are some
error conditions which should not be fatal to the process (such as if
the portal we are trying to discover is not reachable on one of a number
of interfaces), so allow the other ifaces to still be tried in those
cases.

Signed-off-by: Jim Ramsay <jim_ramsay@dell.com>