iscsi tools: add tgt reset to session info and fix unknown values
This patch adds tgt reset to the displayed timeouts.

It also fixes recovery timeout handling. -1 is a valid
setting in newer tools. If the kernel displays it then
it is ok.

And instead of printing "<NULL>" for settings we do not
know the value for this prints out "<empty>" like is done
in other places.

Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>