iscsi tools: fix netlink msg setup
This fixes a regression added in
deef9e20c2f1b107760b5b4c29289de77fcdc39f
and fixes a alignment issue with the new iface netconfig support.

The regression was added when we changed how we sent
the netlink message header and data from one blob to
iovecs. The nlmsg_len was not tracking the header
so the kernel was misreading how much data it had to read.
This would lead to iscsiadm hanging because logins were
not getting completed.

This also fixes a issue where the data was not padded
so when sending multiple params the params would not
start at nice boundaries and could cause unaligned access
errors.