iscsistart: fix iface overriding
The user may only pass in the very basic values when overriding
iface settings, so this patch clears out possibly conflicting
values. For example the user may do

iscsistart -b --param iface.hwaddress=$MAC

but we were leaving the old netdev value in iface.net_ifacename
if it is was in ibft/iscsi_boot.