iscsi tools: fix bnx2i login
This patch fixes a regression added by:

commit 04b4a6699f63c4f0bab9523aae3efb8c909d6587
iscsi tools: have iscsi tools bring up offload net iface

I was using SIOCSIFFLAGS instead of SIOCGIFFLAGS, so if
the interface was up we were not detecting this and resetting
it. This then caused iscsiuio to not be ready and it would
end up failing requests with EAGAIN. In many cases we would
run out of retries before iscsiuio was ready and so the
operation would fail.