iscsi tools: fix unknown param warnings
commit 3d87f013c53c3e229e16c79f5d5fa29a9e3fcef8 added a log
message when we do not match the param name in idbm_rec_update_param.
We cannot do this because when reading iscsid.conf to setup a
record the record type might not support all the settings in
iscsid.conf. We then get a flood of unkown params messages in the
log or stdout.

Fix this by moving the match check to the caller if needed.