iscsiadm: bind ifaces to portals found using isns
Bug and patch from Jayamohan Kallickal at Emulex.

This adds support to be able to use isns discovery and
bind the portals found to offload ifaces. Note that
this does not do iSNS through the offload card. It
uses a normal net connection but portals that are found
will be setup to login using the offload card.

The problem with the previous code was that isns
discovery code was not reading in the iface info before
binding, so the resulting setup was not binding the offload
card to the record.

[Minor cleanup of patch by Mike Christie]
Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>