Wake up to reap children
If we have fork()ed, in addition to exiting poll() if a fd is active, also
wake up once a second.

This should ensure defunct processes are retired in a timely fashion,
while also minimizing wakeups while not needing to reap, i.e. most of the
time.

Change reap_count to an unsigned int, strictly to turn off that part of
readers' brains where they worry it might ever be negative.

Fix a typo in comment.

Signed-off-by: Andy Grover <agrover@redhat.com>